Move subtask validator methods
