+++ app/Model/Subtask.php
@@ -4,11 +4,9 @@ namespace Kanboard\Model;
 
 use PicoDb\Database;
 use Kanboard\Event\SubtaskEvent;
-use SimpleValidator\Validator;
-use SimpleValidator\Validators;
 
 /**
- * Subtask model
+ * Subtask Model
  *
  * @package  model
  * @author   Frederic Guillot
@@ -450,91 +448,5 @@ class Subtask extends Base
                 }
             }
         });
-    }
-
-    /**
-     * Validate creation
-     *
-     * @access public
-     * @param  array   $values           Form values
-     * @return array   $valid, $errors   [0] = Success or not, [1] = List of errors
-     */
-    public function validateCreation(array $values)
-    {
-        $rules = array(
-            new Validators\Required('task_id', t('The task id is required')),
-            new Validators\Required('title', t('The title is required')),
-        );
-
-        $v = new Validator($values, array_merge($rules, $this->commonValidationRules()));
-
-        return array(
-            $v->execute(),
-            $v->getErrors()
-        );
-    }
-
-    /**
-     * Validate modification
-     *
-     * @access public
-     * @param  array   $values           Form values
-     * @return array   $valid, $errors   [0] = Success or not, [1] = List of errors
-     */
-    public function validateModification(array $values)
-    {
-        $rules = array(
-            new Validators\Required('id', t('The subtask id is required')),
-            new Validators\Required('task_id', t('The task id is required')),
-            new Validators\Required('title', t('The title is required')),
-        );
-
-        $v = new Validator($values, array_merge($rules, $this->commonValidationRules()));
-
-        return array(
-            $v->execute(),
-            $v->getErrors()
-        );
-    }
-
-    /**
-     * Validate API modification
-     *
-     * @access public
-     * @param  array   $values           Form values
-     * @return array   $valid, $errors   [0] = Success or not, [1] = List of errors
-     */
-    public function validateApiModification(array $values)
-    {
-        $rules = array(
-            new Validators\Required('id', t('The subtask id is required')),
-            new Validators\Required('task_id', t('The task id is required')),
-        );
-
-        $v = new Validator($values, array_merge($rules, $this->commonValidationRules()));
-
-        return array(
-            $v->execute(),
-            $v->getErrors()
-        );
-    }
-
-    /**
-     * Common validation rules
-     *
-     * @access private
-     * @return array
-     */
-    private function commonValidationRules()
-    {
-        return array(
-            new Validators\Integer('id', t('The subtask id must be an integer')),
-            new Validators\Integer('task_id', t('The task id must be an integer')),
-            new Validators\MaxLength('title', t('The maximum length is %d characters', 255), 255),
-            new Validators\Integer('user_id', t('The user id must be an integer')),
-            new Validators\Integer('status', t('The status must be an integer')),
-            new Validators\Numeric('time_estimated', t('The time must be a numeric value')),
-            new Validators\Numeric('time_spent', t('The time must be a numeric value')),
-        );
     }
 }
